+++ src/App.tsx
@@ -1,17 +1,16 @@
 import React from 'react';
 import { Routes, Route, Navigate, useNavigate } from 'react-router-dom';
 
-import Boards from 'components/pages/boards/Boards';
-import Welcome from 'components/pages/welcome/Welcome';
-import EditProfile from 'components/pages/edit-profile/EditProfile';
+import Boards from 'components/modules/authentication/pages/boards/Boards';
+import EditProfile from 'components/modules/authentication/pages/edit-profile/EditProfile';
 import PrivateRoutes from 'components/private-routes/PrivateRoutes';
 import { Header } from 'components/modules/Header/Header';
 import { useAppDispatch, useAuthUser } from 'hooks/hooks';
 import { removeUserData } from 'store/authorizationSlice';
 import { ToastContainer } from 'react-toastify';
 import { ROUTES, TOASTIFY_SETTINGS } from 'constants/constants';
-import SignIn from 'components/pages/signIn/SignIn';
-import SignUp from 'components/pages/signUp/SignUp';
+import SignIn from 'components/modules/authentication/pages/signIn/SignIn';
+import SignUp from 'components/modules/authentication/pages/signUp/SignUp';
 import { WelcomPage } from 'components/modules/WelcomPage/WelcomPage';
 import { Footer } from 'components/modules/Footer/Footer';
