+++ src/components/modules/Header/ChangeLang/ChangeLang.module.scss
@@ -2,7 +2,7 @@
     width: 2.5rem;
     text-align: center;
     border-radius: 10rem;
-    background-color: rgba(202, 77, 94, 1);
+    background-color: #808080;
     padding: .6rem;
     color: whitesmoke;
     cursor: pointer;
