+++ src/components/modules/Header/Header.tsx
@@ -1,4 +1,5 @@
 import { ROUTES } from 'constants/constants';
+import { useAuthUser } from 'hooks/hooks';
 import React, { useEffect, useState } from 'react';
 import { NavLink } from 'react-router-dom';
 import { Button } from './Button/Button';
@@ -7,6 +8,8 @@ import module from './Header.module.scss';
 
 export const Header = () => {
   const [prevScroll, setScroll] = useState(false);
+  const { userExist } = useAuthUser();
+
   useEffect(() => {
     window.onscroll = () => {
       window.scrollY < 10 ? setScroll(false) : setScroll(true);
@@ -19,23 +22,43 @@ export const Header = () => {
           Project Management System
         </NavLink>{' '}
       </div>
-      <div className={module.wrapper_login_register}>
-        <Button
-          pathLink={ROUTES.signin}
-          borderColor="none"
-          colorText="whitesmoke"
-          title="Войти"
-          fullBackground="rgba(68, 165, 79, 1)"
-        />
-        <Button
-          pathLink={ROUTES.signup}
-          borderColor="none"
-          colorText="whitesmoke"
-          title="Регистрация"
-          fullBackground="rgba(204, 124, 54, 1)"
-        />
-        <ChangeLang />
-      </div>
+      {userExist ? (
+        <div className={module.wrapper_login_register}>
+          <Button
+            pathLink={ROUTES.boards}
+            borderColor="none"
+            colorText="whitesmoke"
+            title="Boards"
+            fullBackground="#0077b5"
+          />
+          <Button
+            pathLink={ROUTES.edit}
+            borderColor="none"
+            colorText="whitesmoke"
+            title="Edit profile"
+            fullBackground="rgba(204, 124, 54, 1)"
+          />
+          <ChangeLang />
+        </div>
+      ) : (
+        <div className={module.wrapper_login_register}>
+          <Button
+            pathLink={ROUTES.signin}
+            borderColor="none"
+            colorText="whitesmoke"
+            title="Sing In"
+            fullBackground="rgba(68, 165, 79, 1)"
+          />
+          <Button
+            pathLink={ROUTES.signup}
+            borderColor="none"
+            colorText="whitesmoke"
+            title="Sing Up"
+            fullBackground="rgba(204, 124, 54, 1)"
+          />
+          <ChangeLang />
+        </div>
+      )}
     </header>
   );
 };
