+++ src/components/modules/authentication/pages/edit-profile/editProfile.module.scss
