+++ src/components/modules/authentication/pages/signUp/SignUp.module.scss
