+++ src/components/modules/authentication/pages/signUp/SignUp.tsx
@@ -1,12 +1,12 @@
 import React, { memo, useState } from 'react';
 import { SubmitHandler, useForm } from 'react-hook-form';
 import styles from './SignUp.module.scss';
-import InputField from 'components/inputField/InputField';
+import InputField from '../../inputField/InputField';
 import { ERROR_TEXT, REGEX, ROUTES, URLS } from 'constants/constants';
-import FormBtn from 'components/formBtn/FormBtn';
+import FormBtn from '../../formBtn/FormBtn';
 import { toast } from 'react-toastify';
 import { updateToast } from 'utils/utils';
-import Preloader from 'components/preloader/Preloader';
+import Preloader from '../../preloader/Preloader';
 
 interface ISignUpData {
   name: string;
