+++ src/components/modules/authentication/pages/welcome/welcome.module.scss
